keybinds: on pkill of anyrun, run anyrun right after
Super+R previously ran the pseudocode
  IF anyrun is running,
    pkill anyrun
  ELSE
    execute anyrun
because anyrun stays alive if the user runs a clipboard copy.

But this means that the user may be confused on whether the next Super+R
should be single or double tapped.

Therefore we now use
  IF anyrun is running,
    pkill anyrun
  execute anyrun
although this means that if anyrun is running but is currently visible
(e.g. user accidentally hit the keybind, and its empty) ESC will be used
to pkill it instead of Super+R.

This is a necessary tradeoff for predictable responsiveness